feat(Chart Item): Add Chart Item Service, Controller and Routes

+++ app/models/CartItemModel.js
@@ -31,7 +31,7 @@ const CartItem = sequelize.define(
          allowNull: false,
       },
       price: {
-         type: DataTypes.DECIMAL(10, 2),
+         type: DataTypes.INTEGER,
          allowNull: false,
       },
    },
